requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-jose[cryptography]==3.5.0
passlib[bcrypt]
bcrypt==3.2.2
python-multipart==0.0.20
email-validator==2.3.0
pydantic==2.11.7
python-dotenv==1.0.0
alembic==1.11.1
pytest==7.4.2
redis==5.0.1
tenacity
google-auth
requests
